+++ plugins/amqp-wakamiti-plugin/src/main/java/module-info.java
@@ -2,7 +2,7 @@ module es.iti.wakamiti.amqp {
 
     requires awaitility;
     requires com.rabbitmq.client;
-    requires imconfig;
     requires iti.commons.jext;
     requires es.iti.wakamiti.api;
+    requires org.apache.commons.lang3;
 }