+++ plugins/cucumber-exporter-wakamiti-plugin/src/main/java/es/iti/wakamiti/plugins/cucumber/CucumberExporterConfig.java
@@ -1,12 +1,12 @@
 package es.iti.wakamiti.plugins.cucumber;
 
-import imconfig.Configuration;
-import imconfig.Configurer;
+import es.iti.wakamiti.api.imconfig.Configuration;
+import es.iti.wakamiti.api.imconfig.Configurer;
 import es.iti.commons.jext.Extension;
 import es.iti.wakamiti.api.extensions.ConfigContributor;
 
 
-@Extension(provider =  "es.iti.wakamiti", name = "cucumber-exporter-config", version = "2.5",
+@Extension(provider =  "es.iti.wakamiti", name = "cucumber-exporter-config", version = "2.6",
     extensionPoint =  "es.iti.wakamiti.api.extensions.ConfigContributor")
 public class CucumberExporterConfig implements ConfigContributor<CucumberExporter> {