+++ wakamiti-api/src/main/java/module-info.java
@@ -6,26 +6,35 @@ module es.iti.wakamiti.api {
 
     exports es.iti.wakamiti.api;
     exports es.iti.wakamiti.api.extensions;
+    exports es.iti.wakamiti.api.imconfig;
     exports es.iti.wakamiti.api.annotations;
     exports es.iti.wakamiti.api.plan;
     exports es.iti.wakamiti.api.util;
+    exports es.iti.wakamiti.api.auth.oauth;
     exports es.iti.wakamiti.api.event;
     exports es.iti.wakamiti.api.datatypes;
     exports es.iti.wakamiti.api.model;
+    exports es.iti.wakamiti.api.matcher;
 
-    requires transitive imconfig;
     requires transitive iti.commons.jext;
     requires transitive org.slf4j;
 
-    requires org.apache.commons.lang3;
     requires org.hamcrest;
     requires slf4jansi;
     requires java.instrument;
     requires java.xml;
+    requires org.apache.commons.configuration2;
+    requires org.apache.commons.lang3;
     requires org.apache.xmlbeans;
+    requires com.fasterxml.jackson.core;
     requires com.fasterxml.jackson.databind;
     requires json.path;
     requires org.apache.groovy;
+    requires org.yaml.snakeyaml;
+    requires java.net.http;
+    requires org.apache.httpcomponents.httpclient;
+    requires org.apache.httpcomponents.httpcore;
+    requires plexus.utils;
 
     uses PropertyEvaluator;
     uses WakamitiAPI;
@@ -38,8 +47,10 @@ module es.iti.wakamiti.api {
     uses Reporter;
     uses StepContributor;
     uses LoaderContributor;
+    uses es.iti.wakamiti.api.imconfig.ConfigurationFactory;
 
     provides ConfigContributor with WakamitiConfiguration;
+    provides es.iti.wakamiti.api.imconfig.ConfigurationFactory with es.iti.wakamiti.api.imconfig.internal.ApacheConfiguration2Factory;
 
     opens es.iti.wakamiti.api.plan to com.fasterxml.jackson.databind;