+++ wakamiti-engine/wakamiti-core/src/main/java/es/iti/wakamiti/core/backend/RunnableBackend.java
@@ -20,7 +20,7 @@ import es.iti.wakamiti.api.util.Pair;
 import es.iti.wakamiti.api.util.ThrowableRunnable;
 import es.iti.wakamiti.core.Wakamiti;
 import es.iti.wakamiti.core.util.LocaleLoader;
-import imconfig.Configuration;
+import es.iti.wakamiti.api.imconfig.Configuration;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 
@@ -40,9 +40,6 @@ import java.util.regex.Matcher;
 public class RunnableBackend extends AbstractBackend {
 
     public static final Logger LOGGER = Wakamiti.LOGGER;
-    public static final String UNNAMED_ARG = "unnamed";
-    public static final String DOCUMENT_ARG = "document";
-    public static final String DATATABLE_ARG = "datatable";
     private static final List<String> DATA_ARG_ALTERNATIVES = List.of(DOCUMENT_ARG, DATATABLE_ARG);
 
     private final PlanNode testCase;
@@ -165,9 +162,12 @@ public class RunnableBackend extends AbstractBackend {
      */
     @Override
     public void setUp() {
+        String type = "set-up";
+        LOGGER.debug("Performing {} operations...", type);
         for (ThrowableRunnable setUpOperation : setUpOperations) {
-            runMethod(setUpOperation, "set-up");
+            runMethod(setUpOperation, type);
         }
+        LOGGER.debug("{} finished", type);
     }
 
     /**
@@ -176,9 +176,12 @@ public class RunnableBackend extends AbstractBackend {
      */
     @Override
     public void tearDown() {
+        String type = "tear-down";
+        LOGGER.debug("Performing {} operations...", type);
         for (ThrowableRunnable tearDownOperation : tearDownOperations) {
-            runMethod(tearDownOperation, "tear-down");
+            runMethod(tearDownOperation, type);
         }
+        LOGGER.debug("{} finished", type);
     }
 
     /**
@@ -190,6 +193,15 @@ public class RunnableBackend extends AbstractBackend {
      */
     private void runMethod(ThrowableRunnable operation, String type) {
         try {
+            Locale locale = LocaleLoader.forLanguage(testCase.language());
+            WakamitiStepRunContext.set(
+                    new WakamitiStepRunContext(
+                            configuration,
+                            this,
+                            locale,
+                            locale
+                    )
+            );
             operation.run();
         } catch (Exception | Error e) {
             Throwable tr = e;
@@ -278,7 +290,8 @@ public class RunnableBackend extends AbstractBackend {
             Map<String, Argument> arguments = stepBackend.invokingArguments();
             step.arguments().addAll(arguments.values());
             Object result = stepBackend.runnableStep().run(arguments);
-            ((List<Object>) extraProperties.get(ContextMap.RESULTS_PROP)).add(result);
+            ((Map<String, Object>) extraProperties.get(ContextMap.RESULTS_PROP))
+                    .put(step.properties().getOrDefault("id", step.id()), result);
             step.prepareExecution().markFinished(clock.instant(), Result.PASSED);
         } catch (Throwable e) {
             fillErrorState(step, instant, e, stepBackend.classifier());
@@ -385,7 +398,7 @@ public class RunnableBackend extends AbstractBackend {
 
         ContextMap() {
             super.put(ID_PROP, testCase.id());
-            super.put(RESULTS_PROP, new LinkedList<>());
+            super.put(RESULTS_PROP, new LinkedHashMap<>());
         }
 
         @Override