+++ wakamiti-engine/wakamiti-core/src/main/java/es/iti/wakamiti/core/datatypes/assertion/BinaryStringAssertProvider.java
@@ -6,7 +6,7 @@
 package es.iti.wakamiti.core.datatypes.assertion;
 
 
-import es.iti.wakamiti.core.backend.ExpressionMatcher;
+import es.iti.wakamiti.api.ExpressionMatcher;
 import org.hamcrest.Matcher;
 import org.hamcrest.Matchers;
 
@@ -114,7 +114,7 @@ public class BinaryStringAssertProvider extends AbstractAssertProvider {
      * {@inheritDoc}
      */
     @Override
-    protected LinkedList<String> regex(Locale locale) {
+    public LinkedList<String> regex(Locale locale) {
         return Arrays.stream(expressions())
                 .map(exp -> ExpressionMatcher.computeRegularExpression(bundle(locale).getString(exp)))
                 .map(exp -> exp.replace(VALUE_WILDCARD, "(\"([^\"\\\\]*(\\\\.[^\"\\\\]*)*)\"|'([^'\\\\]*(\\\\.[^'\\\\]*)*)')"))