+++ wakamiti-engine/wakamiti-core/src/main/java/es/iti/wakamiti/core/runner/PlanNodeLogger.java
@@ -11,7 +11,7 @@ import es.iti.wakamiti.api.model.ExecutionState;
 import es.iti.wakamiti.api.plan.NodeType;
 import es.iti.wakamiti.api.plan.PlanNode;
 import es.iti.wakamiti.api.plan.Result;
-import imconfig.Configuration;
+import es.iti.wakamiti.api.imconfig.Configuration;
 import org.slf4j.Logger;
 
 import java.time.Duration;