+++ wakamiti-engine/wakamiti-maven-plugin/src/main/java/es/iti/wakamiti/maven/WakamitiConfigurable.java
@@ -7,8 +7,8 @@ package es.iti.wakamiti.maven;
 
 
 import es.iti.wakamiti.core.Wakamiti;
-import imconfig.Configuration;
-import imconfig.ConfigurationException;
+import es.iti.wakamiti.api.imconfig.Configuration;
+import es.iti.wakamiti.api.imconfig.ConfigurationException;
 import org.apache.maven.plugin.logging.Log;
 
 import java.nio.file.Path;